% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_request.r
\name{nsc_config}
\alias{nsc_config}
\title{Create a National Student Clearinghouse configuration object}
\usage{
nsc_config(schoolCode, branchCode, schoolName)
}
\arguments{
\item{schoolCode}{NSC school code}

\item{branchCode}{NSC branch code}

\item{schoolName}{NSC school name}
}
\value{
A dataframe configuration for passing to nsc_request* calls.
}
\description{
Create a National Student Clearinghouse configuration object for
passing to nsc_request* calls.
}
\details{
Create a National Student Clearinghouse configuration object. All fields are strings.

\if{html}{\out{<div class="sourceCode">}}\preformatted{nsc_conf <- nsc_config(
                schoolCode="YOURFICECODE",
                branchCode="YOURBRANCHCODE",
                schoolName="YOURSCHOOLNAME"
            )
}\if{html}{\out{</div>}}
}
\seealso{
\url{https://studentclearinghouse.info/onestop/wp-content/uploads/STCU_User_Manual.pdf} for
details on how the request file is supposed to be structured.
}
