% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_request.r
\name{nsc_request}
\alias{nsc_request}
\title{Create a National Student Clearinghouse request file}
\usage{
nsc_request(
  df,
  config,
  inquiryType,
  path,
  fn,
  search,
  enrolledStudents,
  write = TRUE
)
}
\arguments{
\item{df}{Dataframe to convert into format for submission to NSC}

\item{config}{A configuration object for NSC files}

\item{inquiryType}{One of CO (Longitudinal Cohort), DA (Declined Admission),
PA (Prior Attendance), or SE (Subsequent Enrollment)}

\item{path}{Path where NSC file will be saved (default=\code{getwd()})}

\item{fn}{Name of the NSC file (default=same name as the dataframe with "_{inquiryType}.tsv" appended)}

\item{search}{If the dataframe does not include a 'Search Begin Date' field, this will be used (default=TODAY)}

\item{enrolledStudents}{Are these students currently enrolled? Set to FALSE
for PA query to allow SSN. (default=FALSE for SE, otherwise TRUE)}

\item{write}{Whether function should write out the file}
}
\value{
This function is called mainly for its side-effects (writing out the
request file) but also returns the student records used to create the
request file.
}
\description{
Create a National Student Clearinghouse request file in the required format.
}
\details{
This requires a configuration object which contains required elements.

\if{html}{\out{<div class="sourceCode">}}\preformatted{nsc_conf <- nsc_config(
                schoolCode=YOURFICECODE,
                branchCode=YOURBRANCHCODE,
                schoolName=YOURSCHOOLNAME
            )
}\if{html}{\out{</div>}}

DOB must be either a date or a character string with format \verb{\%Y\%m\%d}.
}
\seealso{
\url{https://studentclearinghouse.info/onestop/wp-content/uploads/STCU_User_Manual.pdf} for
details on how the request file is supposed to be structured.
}
