% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_request.r
\name{nsc_config}
\alias{nsc_config}
\title{Create a National Student Clearinghouse configuration object}
\usage{
nsc_config(schoolCode, branchCode, schoolName)
}
\arguments{
\item{schoolCode}{NSC school code}

\item{branchCode}{NSC branch code}

\item{schoolName}{NSC school name}
}
\description{
Create a National Student Clearinghouse configuration object. All fields are strings.
nsc_conf <- nsc_config(
schoolCode="YOURFICECODE",
branchCode="YOURBRANCHCODE",
schoolName="YOURSCHOOLNAME"
)
}
