% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_request.r
\name{nsc_request}
\alias{nsc_request}
\title{Create a National Student Clearinghouse request file}
\usage{
nsc_request(
  df,
  config,
  inquiryType = "SE",
  path = getwd(),
  fn = paste0(deparse(substitute(df)), "_SE.tsv"),
  search = format(Sys.Date(), "\%Y\%m\%d"),
  enrolledStudents = ifelse(inquiryType == "SE", FALSE, TRUE)
)
}
\arguments{
\item{df}{Dataframe to convert into format for submission to NSC}

\item{config}{A configuration dataframe for NSC files}

\item{inquiryType}{One of CO (Longitudinal Cohort), DA (Declined Admission), PA (Prior Attendance), or SE (Subsequent Enrollment)}

\item{path}{Path where NSC file will be saved (default=getwd())}

\item{fn}{Name of the NSC file (default=same name as the dataframe with "_SE.tsv" appended)}

\item{search}{If the dataframe does not include a 'Search Begin Date' field, this will be used (default=TODAY)}

\item{enrolledStudents}{Are these students currently enrolled? Set to FALSE for PA query to allow SSN.}
}
\description{
Create a National Student Clearinghouse request file in the required format.
This requires a config file which contains required elements.
DOB must be either a date or a character string with format "%Y%m%d".
}
