% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_request.r
\name{nsc_request_se}
\alias{nsc_request_se}
\title{Create a National Student Clearinghouse Subsequent Enrollment (SE) request file}
\usage{
nsc_request_se(df, config, path, fn, search, write = TRUE)
}
\arguments{
\item{df}{Dataframe to convert into format for submission to NSC}

\item{config}{A configuration object for NSC files}

\item{path}{Path where NSC file will be saved (default=\code{getwd()})}

\item{fn}{Name of the NSC file (default=same name as the dataframe with "_{inquiryType}.tsv" appended)}

\item{search}{If the dataframe does not include a 'Search Begin Date' field, this will be used (default=TODAY)}

\item{write}{Whether function should write out the file}
}
\description{
Create a National Student Clearinghouse Subsequent Enrollment (SE) request file
in the required format.
This requires a configuration object which contains required elements.
nsc_conf <- nsc_config(
schoolCode="YOURFICECODE",
branchCode="YOURBRANCHCODE",
schoolName="YOURSCHOOLNAME"
)
DOB must be either a date or a character string with format "\%Y\%m\%d".
}
