% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_return_se_convert.r
\name{nsc_return_se_convert}
\alias{nsc_return_se_convert}
\title{Convert National Student Clearinghouse return file to database format}
\usage{
nsc_return_se_convert(fn)
}
\arguments{
\item{fn}{Full path to the details file provided by NSC}

\item{type}{The type of merge. Possible options are csv, fixed, other}
}
\description{
Convert the National Student Clearinghouse return file to a database format
that is easier to work with. Each student will have one record per college
attended. Additional records are added for each degree earned. If a student
did not attend a college that reports to the NSC, then Record Found Y/N will
be 'N'.
}
\keyword{file}
